# web/static/openapi.yaml
openapi: 3.0.3
info:
  title: Chatwoot Lite WhatsApp Gateway API
  version: 1.1.0
  description: |
    API para consultar conversas e enviar mensagens pelo WhatsApp Cloud API.
    No ambiente de testes, clique em Authorize e informe o API_ACCESS_TOKEN.
servers:
  - url: /
    description: Ambiente atual
security:
  - ApiAccessToken: []
...
paths:
...
  /api/config/meta:
    get:
      tags: [Configuração Meta]
      summary: Consulta a configuração Meta ativa
      security: []
      responses:
        '200': { description: Configuração atual }
    post:
      tags: [Configuração Meta]
      summary: Salva credenciais da Meta
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number_id, access_token]
              properties:
                phone_number_id: { type: string }
                phone_number: { type: string }
                access_token: { type: string, format: password }
                api_version: { type: string, example: v23.0 }
      responses:
        '200': { description: Credenciais salvas }
        '500': { description: Falha ao persistir credenciais }
components:
  securitySchemes:
    ApiAccessToken:
      type: apiKey
      in: header
      name: api_access_token
      description: Valor configurado em API_ACCESS_TOKEN.
...